thesis: update bibliography
I'm now generating this from Zotero, which is why the citation key for
gauck_implementing_2022 has changed. I may change it back later.

Also TBD is how to cite the BEPs. I'd still like inline citations to be
"BEP 00xx" instead of the author's names or the titles, but I haven't
figured that out yet and it's not high priority. Notice that picking the
"Standard" item type in Zotero has produced a `misc` entry in biblatex.

If I have time, I also want to reinvestigate using Pandoc's citation
processing. The original issue was that I couldn't get the bibliography
in the right place, but there's a section in the Pandoc manual that
might let me place the bibliography where I want and still use citeproc,
which I think would be nicer. See:
<https://pandoc.org/MANUAL.html#placement-of-the-bibliography>