thesis: set footer at correct location
Because I was using `before-body`, the footer was defined before the
frontmatter (including prelims and table of contents). Because each
prelim counts as a chapter, the footer on each of those pages is
overridden to be blank (`\thispagestyle{empty}`). However, since my
table of contents now overflows onto a second page, I was seeing the
footer appear there. I now properly only enable the footer at the end of
the frontmatter and before the mainmatter, avoiding this bug.