template: add option for folio binding
Currently, this changes the paper size to have the 11x17 aspect ratio
(important since I am printing two pages to a sheet on letter paper).
I think I may also adjust the margins when printing for a different
size, although this may be unnecessary now that LaTeX's view of the
paper is 8.5x13.136 instead of 7.117x11.